Inline types moved to new namespace + fix phpcs bug
